pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "reacmd"
authors = [{name = "Joe Stanley", email = "engineerjoe440@yahoo.com"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "python-reapy"
]

[project.urls]
Home = "https://github.com/engineerjoe440/ReaCmd"

[project.scripts]
rea-record = "reacmd.record:entrypoint"

